refactor: create timeline component, richtext hook
