+++ frontend/hooks/richText.tsx
@@ -1,0 +1,31 @@
+import groq from "groq";
+import client from "../client";
+import imageUrlBuilder from "@sanity/image-url";
+
+function urlFor(source: any) {
+  return imageUrlBuilder(client).image(source);
+}
+
+const ptComponents = {
+  types: {
+    image: ({ value }: any) => {
+      if (!value?.asset?._ref) {
+        return null;
+      }
+      return (
+        <img
+          alt={value.alt || " "}
+          loading="lazy"
+          src={urlFor(value)
+            .width(320)
+            .height(240)
+            .fit("max")
+            .auto("format")
+            .url()}
+        />
+      );
+    },
+  },
+};
+
+export default ptComponents;
